eleventh commit

+++ src/main/resources/static/movies.js
@@ -14,6 +14,32 @@ $(document).ready(()=>{
         const year = $('#modal-movie-year').val();
         const description = $('#modal-movie-description').val();
         const rating = $('#modal-movie-rating').val();
+        const poster1 = $('#modal-movie-poster1').val();
+        const poster2 = $('#modal-movie-poster2').val();
+        const poster3 = $('#modal-movie-poster3').val();
+        const poster4 = $('#modal-movie-poster4').val();
+        const trailer1 = $('#modal-movie-trailer1').val();
+        const trailer2 = $('#modal-movie-trailer2').val();
+        const posters = [];
+            if(poster1 !== ''){
+                posters.push(poster1);
+            }
+            if(poster2 !== ''){
+                posters.push(poster2);
+            }
+            if(poster3 !== ''){
+                posters.push(poster3);
+            }
+            if(poster4 !== ''){
+                posters.push(poster4);
+            }
+        const trailers = [];
+            if(trailer1 !== ''){
+                trailers.push(trailer1);
+            }
+            if(trailer2 !== ''){
+                trailers.push(trailer2);
+            }
 
         fetch(`/api/movies/${movieToEdit}`, {
             method: 'PUT',
@@ -21,7 +47,9 @@ $(document).ready(()=>{
                 name: name,
                 year: year,
                 description: description,
-                rating: rating
+                rating: rating,
+                posters: posters,
+                trailers: trailers
             }),
             headers: {
                 'Content-Type': 'application/json'
@@ -40,6 +68,32 @@ $(document).ready(()=>{
         const year = $('#modal-movie-year').val();
         const description = $('#modal-movie-description').val();
         const rating = $('#modal-movie-rating').val();
+        const poster1 = $('#modal-movie-poster1').val();
+        const poster2 = $('#modal-movie-poster2').val();
+        const poster3 = $('#modal-movie-poster3').val();
+        const poster4 = $('#modal-movie-poster4').val();
+        const trailer1 = $('#modal-movie-trailer1').val();
+        const trailer2 = $('#modal-movie-trailer2').val();
+        const posters = [];
+            if(poster1 !== ''){
+                posters.push(poster1);
+            }
+            if(poster2 !== ''){
+                posters.push(poster2);
+            }
+            if(poster3 !== ''){
+                posters.push(poster3);
+            }
+            if(poster4 !== ''){
+                posters.push(poster4);
+            }
+        const trailers = [];
+            if(trailer1 !== ''){
+                trailers.push(trailer1);
+            }
+            if(trailer2 !== ''){
+                trailers.push(trailer2);
+            }
 
         fetch('/api/movies', {
             method: 'POST',
@@ -47,7 +101,10 @@ $(document).ready(()=>{
                 name: name,
                 year: year,
                 description: description,
-                rating: rating
+                rating: rating,
+                posters: posters,
+                trailers: trailers
+
             }),
             headers: {
                 'Content-Type': 'application/json'
@@ -67,9 +124,19 @@ $(document).ready(()=>{
         $('#modal-movie-year').val('');
         $('#modal-movie-description').val('');
         $('#modal-movie-rating').val('');
+        $('#modal-movie-poster1').val('');
+        $('#modal-movie-poster2').val('');
+        $('#modal-movie-poster3').val('');
+        $('#modal-movie-poster4').val('');
+        $('#modal-movie-trailer1').val('');
+        $('#modal-movie-trailer2').val('');
     }
 
     $('.delete-icon').click(function () {
+        let confirm = window.confirm("Are you sure you want to delete the entry?");
+        if (!confirm) {
+            return false;
+        }
         const movieId = this.parentElement.id;
         fetch(`/api/movies/${movieId}`, {
             method: 'DELETE'
@@ -88,6 +155,12 @@ $(document).ready(()=>{
         $('#modal-movie-year').val(year);
         $('#modal-movie-description').val(description);
         $('#modal-movie-rating').val(rating);
+        $('#modal-movie-poster1').val(poster.length > 0 ? poster[0].url : '');
+        $('#modal-movie-poster2').val(poster.length > 1 ? poster[1].url : '');
+        $('#modal-movie-poster3').val(poster.length > 2 ? poster[2].url : '');
+        $('#modal-movie-poster4').val(poster.length > 3 ? poster[3].url : '');
+        $('#modal-movie-trailer1').val(trailer.length > 0 ? trailer[0].url : '');
+        $('#modal-movie-trailer2').val(trailer.length > 1 ? trailer[1].url : '');
     });
 
     $('#add-movie-main-button').click(() => {
